// synthesis-web/src/app/api/check.ts
export enum CheckEnum {
    NO_TEXT = 'no_text',
    MIN_WORDS = 'min_words',
    MAX_WORDS = 'max_words',
    NO_NUMBERS = 'no_numbers',
    NO_ABBREVIATIONS = 'no_abbreviations',
    LITHUANIAN = 'lithuanian',
    PROFANITY = 'profanity'
}

export enum ErrorCodes {
    UNAUTHORIZED = 'unauthorized',
    OUT_OF_QUOTA = 'out_of_quota',
}